Move details panel component to shared location
