+++ src/components/tables/Details/DetailsPanel.tsx
